Convert to Specialised Age Groups

Modify the age groups used by labels. The new age groups must contain the old age groups, and follow a regular pattern:

age_modify_five Five-year age groups

age_modify_ten Ten-year age groups

age_modify_life Age groups used in 'abridged' life tables

Usage

age_modify_five(labels, interpret_fail = c("error", "warn", "silent"))

age_modify_ten(labels, interpret_fail = c("error", "warn", "silent"))

age_modify_life(labels, interpret_fail = c("error", "warn", "silent"))

Arguments

labels

Vector of age group labels.

interpret_fail

Action if element of labels cannot be interpreted. Choices are "error" (the default), "warn", and "silent".

Value

Character vector or factor with the same length as labels.

Abridged and complete life tables

An 'abridged' life table uses age groups "0", "1-4", "5-9", "10-14", and so on up to the oldest age group, which is open on the right. A 'complete' life table uses single-year age groups.

See also

Examples

labels <- c("1-3", "87-89", "0", "91+", "total", "52")
age_modify_five(labels)
#> [1] "0-4"   "85-89" "0-4"   "90+"   "Total" "50-54"
age_modify_ten(labels)
#> [1] "0-9"   "80-89" "0-9"   "90+"   "Total" "50-59"
age_modify_life(labels)
#> [1] "1-4"   "85-89" "0"     "90+"   "Total" "50-54"